requirements.txt
tensorflow==1.15
gpt_2_simple
python-telegram-bot
Pillow
selenium
feedparser
